build: Make it less confusing for themers
Bring back the old assets directory. With this commit the way it
encrypt the assets is changed too.

First, it makes a temporary assets directory  in 'assets-temp', then it
will encrypt all the overlays file inside 'assets' directory. After the
build finish, it deletes encrypted files in 'assets' and copies back
all the temporary assets directory to 'assets'